requirements.txt
selenium==4.12.0
webdriver_manager==4.0.0
rq==1.10.0
redis==3.5.3
python-dotenv==0.19.1 
httpx==0.19.0 
